requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai-whisper==20231117
ffmpeg-python==0.2.0
pydub==0.25.1
google-generativeai==0.4.0
transformers==4.35.0
torch==2.1.0
numpy==1.24.3
scikit-learn==1.3.2
accelerate==0.24.1
bitsandbytes==0.41.3
einops==0.7.0

huggingface-hub==0.17.1
werkzeug==3.0.1

